requirements.txt
fastapi
uvicorn
pydantic-settings
bcrypt
motor
email-validator